fix rendering widget
