+++ src/main.tsx
@@ -1,9 +1,18 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import App from "./App.tsx";
+import Widget from "./widget/index.tsx";
 
-createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>
-);
+if (Boolean(import.meta.env.DEV)) {
+  const testRootElementId = "test-root-element";
+  const isRootElementTest = import.meta.env.VITE_TEST_ROOT_ELEMENT === "true";
+  if (isRootElementTest) {
+    const testRootElement = document.createElement("div");
+    testRootElement.id = testRootElementId;
+    document.body.appendChild(testRootElement);
+  }
+
+  new Widget({
+    apiKey: "b7511851-0a8b-4ee4-b14c-09e33d453cfd",
+    localServerBaseUrl: "http://localhost:3001",
+  });
+}
+
+window.RahuiWidget = Widget;
